test(members): guard the seat-change reset, split the suite, and match the new success shape
Three CI failures, all from this branch:

- The per-test reset used a raw deleteMany filtered on ids assigned in
  beforeAll, so a setup that threw would have matched every role binding in
  the shared database. It goes through cleanupTestRows now, which refuses an
  unidentified filter instead. The teardownScan gate caught it.
- The dialog's mocked updateMemberRole resolved undefined while the real one
  now answers with the teams left without a team admin, so the save read a
  field off nothing and never reached the bindings mutation. The mock answers
  what the mutation answers. The component defaults the field too: mid-rollout
  this code can reach a server that omits it, and by then the save has already
  succeeded, so reading it eagerly would report a successful save as failed.
- The plan-limit suite asserted the success shape exactly, which the new field
  changes.

The seat-change suite also passed 300 source lines, so it splits along the line
the behaviour already had: what a seat correction may do, and what a team-local
decision still may not, over one shared fixture.
